Remove `variable-transfers-main` and add `overview-networks`
Updated based on Jay's feedback
